Initial version (WIP) (#1)
This Notebook provides Python functions to facilitate working with Gen3 data in Terra.

This is Notebook is currently a **work in progress** and is not yet intended for general availability/use.  
Some of the more important known limitations are:
* Subjects are only included in the resulting consolidated table if there is an entry corresponding to the subject for all tables in the consolidation/merge specification. There is not yet support for the distinction between required and optional tables.
* The `reference_file` table data is not currently included in the consolidated table. Attempting to do so resulted in much duplication in the resulting table. This may be solvable, yet requires investigtion.
* Provide user-friendly/useful error messages when problems occur.
* (Currently) fails to import "large" (e.g., 8,000+ rows) consolidated tables into Terra
* Needs documentation for end-user use, customization and selected implementation details.
* Remove numeric index from resulting dataframe
* Needs automated tests
* ...

That said, it generally seems be working pretty well and should enable continuing development of example workspaces for GWAS and other analysis. Yet, given its early state, some issues should be expected to be found, and reported when they are.
Continuing evoltion depends on feedback, which is welcomed and encouraged.