Added support for putting the consolidated table into a different workspace
Added support for  having the source data tables (exported from Gen3) in one
project/workspace and putting the resulting consolidated data table in a different
project/workspace.

There are several benefits to this.
It can help to mitigate Terra data table scalability issues.
Having both the source data tables (from Gen3) and the consolidated table
in the same workspace can double the amount of data in that workspace,
and cause some firecloud API operations to fail because they take longer
then the 1-minute firecloud server-side timeout.

This can also be helpful in organizing and managing data.
For example, a researcher may work with multiple synthetic cohorts,
with each cohort in its own workspace, and the consolidated table for
each placed in the researcher's main/primary workspace.
This provides the research with convenient access to the different
cohorts without having the clutter of the original Gen3 data in that same
workspace.

This can also help deal with the large duration required to delete Terra
data tables. By having all the Gen3 data in a different workspace, it may
often be faster/easier to just delete the whole workspace for that cohort
than to delete each of the Gen3 tables individually.

To specify an alternate project/workspace for the consolidated table,
pass the following parameters as keyword arguments to the top-level
table consolidation function:
```
dest_project=my_destination_project, dest_workspace=my_destination_workspace
```
where my_destination_project and my_destination_workspace represent the actual alternate destination
project and workspace names.