Increased number of retry attempts and duration between them
It seems that with large data tables (~40k rows) and perhaps many of them
initial attempts to perform fapi.list_entity_types and fapi.get_entities_tsv
typically timeout. Yet, after retrying multiple times, with enough duration
between them, they eventually complete successfully, then can be
accessed multiple times successfully.
This suggests that enough time is needed for the Rawls database to
move the data from disk into some form of cache (disk or database),
and once that is done, the operation can be completed without the
timeout period and this remains to be the case for as long as the
data is in the cache.
Could/should verify this with the developers, yet this seems like
it may be what is happening here.